requirements.txt
feedparser
python-telegram-bot
python-dotenv
